+++ lib/fields/watchFields.d.ts
@@ -23,6 +23,31 @@ export declare const CREATE_WATCH_FIELDS: ({
     required: boolean;
     helpText: string;
 })[];
+export declare const TEST_WATCH_FIELDS: ({
+    key: string;
+    label: string;
+    type: "string";
+    required: boolean;
+    helpText: string;
+} | {
+    key: string;
+    label: string;
+    type: "text";
+    required: boolean;
+    helpText: string;
+} | {
+    key: string;
+    label: string;
+    type: "integer";
+    required: boolean;
+    helpText: string;
+} | {
+    key: string;
+    label: string;
+    type: "boolean";
+    required: boolean;
+    helpText: string;
+})[];
 export declare const UPDATE_WATCH_FIELDS: ({
     key: string;
     label: string;